Command overhaul
remove cluster command
merge upgrade and up commands
rename down command
